Add Swedish comment about device removal bug
A comment was added to the `DeviceDetailViewModel` class in the
`DeviceDetailViewModel.cs` file. The comment, written in Swedish,
mentions fixing a bug related to the removal of a device that no
longer exists. The comment translates to "FIX WHEN REMOVING IF A
DEVICE NO LONGER EXISTS, BUG," serving as a reminder for future
development to address this issue.