Add moq package and MVVM\ViewModels\ folder structure
Added the `moq` package (v4.20.72) to `SmartHomeMauiApp.Tests.csproj`
to facilitate the creation of mock objects in unit tests. This helps
in isolating the code being tested and improving test reliability.

Included a new `ItemGroup` for the `MVVM\ViewModels\` folder,
indicating an organizational change to better manage ViewModel files
in an MVVM architecture.