Add dynamic response message color to SettingsPage
Updated SettingsViewModel to include a new observable property `ResponseMessageColor` with a default value of "Red". Modified `SaveSettingsAsync` to set `ResponseMessageColor` to "Green" on success and "Red" on error. Updated `SettingsPage.xaml` to bind the `Label`'s `TextColor` to `ResponseMessageColor`. Refactored `SettingsPage` constructor for proper initialization and added `OnAppearing` method to reset response properties.